+++ sumologic/sumologic_polling_source.go
@@ -39,19 +39,25 @@ type PollingAuthentication struct {
 	TokenUrl                string `json:"token_uri"`
 	AuthProviderX509CertUrl string `json:"auth_provider_x509_cert_url"`
 	ClientX509CertUrl       string `json:"client_x509_cert_url"`
+	SharedAccessPolicyName  string `json:"sharedAccessPolicyName"`
+	SharedAccessPolicyKey   string `json:"sharedAccessPolicyKey"`
 }
 
 type PollingPath struct {
-	Type                      string                           `json:"type"`
-	BucketName                string                           `json:"bucketName,omitempty"`
-	PathExpression            string                           `json:"pathExpression,omitempty"`
-	LimitToRegions            []string                         `json:"limitToRegions,omitempty"`
-	LimitToNamespaces         []string                         `json:"limitToNamespaces,omitempty"`
-	LimitToServices           []string                         `json:"limitToServices,omitempty"`
-	CustomServices            []string                         `json:"customServices,omitempty"`
-	TagFilters                []TagFilter                      `json:"tagFilters,omitempty"`
-	SnsTopicOrSubscriptionArn PollingSnsTopicOrSubscriptionArn `json:"snsTopicOrSubscriptionArn,omitempty"`
-	UseVersionedApi           *bool                            `json:"useVersionedApi,omitempty"`
+	Type                      string                            `json:"type"`
+	BucketName                string                            `json:"bucketName,omitempty"`
+	PathExpression            string                            `json:"pathExpression,omitempty"`
+	LimitToRegions            []string                          `json:"limitToRegions,omitempty"`
+	LimitToNamespaces         []string                          `json:"limitToNamespaces,omitempty"`
+	LimitToServices           []string                          `json:"limitToServices,omitempty"`
+	CustomServices            []string                          `json:"customServices,omitempty"`
+	TagFilters                []TagFilter                       `json:"tagFilters,omitempty"`
+	SnsTopicOrSubscriptionArn *PollingSnsTopicOrSubscriptionArn `json:"snsTopicOrSubscriptionArn,omitempty"`
+	UseVersionedApi           *bool                             `json:"useVersionedApi,omitempty"`
+	Namespace                 string                            `json:"namespace,omitempty"`
+	EventHubName              string                            `json:"eventHubName,omitempty"`
+	ConsumerGroup             string                            `json:"consumerGroup,omitempty"`
+	Region                    string                            `json:"region,omitempty"`
 }
 
 type TagFilter struct {